This commit adds a number of useful automagical OTR enhancements
It adds the following configuration options:

  OTRAutoTearDown
  OTRAutoAppendTag
  OTRAutoStartSession

When OTRAutoTearDown is true, we tear down finished OTR sessions.

When OTRAutoAppendTag is true, we tag all of our plaintext messages
with the OTRv2 whitespace-tag specified in the OTR protocol

When OTRAutoStartSession is true, we attempt to automatically start
OTR sessions with people who appear to support OTRv2 and higher.