Consolidate the well known Hidden Services.
Having the repeated code looked a bit weird, so pulling it all into a map and
doing it that way seemed a bit cleaner to me. Also, as a bonus it's easier to
add more "well known" hidden services later.

This assumes that they are always going to be running on port 5222, is this a
reasonable assumption?