Fix the password prompt.
Changes to the terminal package broke the password prompt. When the
terminal was resized it printed the current prompt and, because a prompt
had already been printed, a second prompt was suppressed when asking for
the password.

This change sets the prompt to be empty until after the password has
been read.

Fixes #63.