Support XMPP forms for registration.
When creating accounts in the new style, one has to be careful that the
username and password given to xmpp-client are the same as given in the
form. There doesn't appear to be any magic way around this other than
trying to guess at which fields in the form are the username and
password. This code tries to do that, but it's only tested to work
against jabber.ccc.de and it's very fragile.

Fixes #43.