Pass a custom TLS Configuration
Pass a custom `tls.Config` to the `xmpp` library to improve the security of the
connection.

The two things that the custom configuration does are:
  * Set the minimum TLS version to TLS 1.0 (as opposed to SSLv3)
  * Sets a custom set of CipherSuites which:
    * Require forward secrecy
    * Do not use RC4

Note: This could potentially break some people's connection if their XMPP server
only supports SSLv3 or does not support any of the cipher suites chosen.

This patch depends on agl/xmpp@1e966e3