Trust only CACert for jabber.ccc.de.
The server jabber.ccc.de is popular but uses CACert issued certificates.
Distros are removing this root from their root sets and thus doing the
obvious thing with xmpp-client results in a certificate error.

This change embeds the CACert root certificate and sets the set of
trusted roots to just CACert when the server is "jabber.ccc.de".

Fixes #84