Call VerifyHostname on certificate directly.
Go 1.5 changes the semantics of crypto/tls.Conn.VerifyHostname to reject
the call if certificates weren't verified during the handshake. Because
we need to handle a bunch of extra cases, xmpp-client does the
certificate verification itself so this Go 1.5 broke us when we tried to
verify the hostname.

This change calls VerifyHostname on the leaf certificate directly, which
is all that crypto/tls.Conn.VerifyHostname really did anyway.

Fixes #94