Add I/O timeouts.
Currently, if a connection dies quietly, for instance because a laptop was
suspended, xmpp-client won't notice, and will appear to still be connected.

This change adds one-minute pings and a three-minute timeout on the net.Conn. If
the timeout is exceeded, xmpp-client will exit with an error, offering an
opportunity to restart (either manually or automatically).